LibWeb: Split FontStyleValue out of StyleValue.{h,cpp}
